♻️refactor to reduce code duplication
