+++ src/CMakeLists.txt
@@ -1,10 +1,13 @@
 add_library(${PROJECT_NAME}
             ${CMAKE_CURRENT_SOURCE_DIR}/algorithms/synthesis/syrec_synthesis.cpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/algorithms/synthesis/syrec_synthesis_additional_lines.cpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/algorithms/synthesis/syrec_synthesis_no_additional_lines.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/algorithms/simulation/simple_simulation.cpp
 
             ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/simulation/simple_simulation.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/syrec_synthesis.hpp
-			${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/derived_syrec_synthesis.hpp
+            ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/syrec_synthesis_with_additional_lines.hpp
+            ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/syrec_synthesis_without_additional_lines.hpp
 
             ${CMAKE_CURRENT_SOURCE_DIR}/core/syrec/parser.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/core/syrec/program.cpp
